+++ MoreGrassStarters/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Custom grass now spreads correctly.
+
 ## 1.0.6
 Released 11 September 2021 for SMAPI 3.12.5 or later. Updated by Pathoschild.